requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.0.1
Faker==28.4.1
gunicorn==23.0.0
idna==3.8
ipython==8.26.0
jedi==0.19.1
k==0.0.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
djangorestframework-simplejwt
